+++ glide2gl/src/Glide64/ucode00.h
@@ -581,7 +581,7 @@ static void uc0_moveword(uint32_t w0, uint32_t w1)
       case 0x04:
          if (((w0 >> 8)&0xFFFF) == 0x04)
          {
-            rdp.clip_ratio = vi_integer_sqrt(w1);
+            rdp.clip_ratio = (float)vi_integer_sqrt(w1);
             rdp.update |= UPDATE_VIEWPORT;
          }
          //FRDP ("clip %08lx, %08lx\n", w0, w1);
