+++ glide2gl/src/Glide64/ucode08.h
@@ -244,7 +244,7 @@ static void uc8_moveword(uint32_t w0, uint32_t w1)
       case G_MW_CLIP:
          if (offset == 0x04)
          {
-            rdp.clip_ratio = vi_integer_sqrt(w1);
+            rdp.clip_ratio = (float)vi_integer_sqrt(w1);
             rdp.update |= UPDATE_VIEWPORT;
          }
          break;
